+++ packages/components-js/tests/vrt/playwright/specs/states/tag-dismissible.vrt.ts
@@ -2,7 +2,7 @@ import { expect, type Page, test } from '@playwright/test';
 import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 import {
   forceFocusHoverState,
-  forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -25,8 +25,8 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
   });
 
   await forceHoverState(page, '.hover p-tag-dismissible >>> button');
-  await forceFocusState(page, '.focus p-tag-dismissible'); // native outline should not be visible
-  await forceFocusState(page, '.focus p-tag-dismissible >>> button');
+  await forceFocusVisibleState(page, '.focus p-tag-dismissible'); // native outline should not be visible
+  await forceFocusVisibleState(page, '.focus p-tag-dismissible >>> button');
   await forceFocusHoverState(page, '.focus-hover p-tag-dismissible >>> button');
 };