+++ packages/components-js/tests/vrt/playwright/specs/states/toast-item.vrt.ts
@@ -2,7 +2,7 @@ import { expect, type Page, test } from '@playwright/test';
 import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 import {
   forceFocusHoverState,
-  forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -22,7 +22,7 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
   });
 
   await forceHoverState(page, '.hover p-toast-item >>> p-button-pure >>> button');
-  await forceFocusState(page, '.focus p-toast-item >>> p-button-pure >>> button');
+  await forceFocusVisibleState(page, '.focus p-toast-item >>> p-button-pure >>> button');
   await forceFocusHoverState(page, '.focus-hover p-toast-item >>> p-button-pure >>> button');
 };